Mike Crowe: According to the libcurl documentation the parameter to curl_version_info should be CURLVERSION_NOW rather than a bit value from the features bitmask.
Also, on my installation at least libssh_version is NULL so the check
segfaults. It would seem that ssl_version is a more appropriate field to
check.

Finally, if libcurl lacks SSL support them ssl_version may be NULL. Make sure we
don't segfault inside strcasestr if that is the case.



git-svn-id: https://gnunet.org/svn/libmicrohttpd@22116 140774ce-b5e7-0310-ab8b-a85725594a96